refactor: move sharing ops to sharing feature; error states to app-level
Sharing operations (invite, collaborate, public/private, role changes)
moved from documents/lib/documents.ts to sharing/lib/document-sharing.ts.
Error state components moved to app/components/error-states.tsx.
ESLint rule tightened so lib/ files cannot cross-feature into widgets/screens.
