+++ src/app/features/sharing/widgets/share-dialog.tsx
@@ -36,7 +36,7 @@ import {
 	isDocumentPublic,
 	getPublicLink,
 	type Collaborator,
-} from "@/app/features/documents"
+} from "../lib/document-sharing"
 import { testIds } from "@/app/lib/test-ids"
 
 export { ShareDialog }
